ref: move fake producer into its own package
