+++ producers/producers.go
@@ -4,14 +4,9 @@ import (
 	"fmt"
 
 	"github.com/zalando-incubator/mate/pkg"
+	"github.com/zalando-incubator/mate/producers/fake"
 	"github.com/zalando-incubator/mate/producers/kubernetes"
 )
-
-var params struct {
-	dnsName      string
-	mode         string
-	targetDomain string
-}
 
 type Producer interface {
 	Endpoints() ([]*pkg.Endpoint, error)
@@ -24,7 +19,7 @@ func New(name string) (Producer, error) {
 	case "kubernetes":
 		return kubernetes.NewProducer()
 	case "fake":
-		return NewFake()
+		return fake.NewFake()
 	}
 	return nil, fmt.Errorf("Unknown producer '%s'.", name)
 }
